requirements.txt
flask
flask-cors
sqlalchemy
python-dotenv
requests
mysql-connector-python